Added jwt token maintenecance, most updated version
JWT accesskey and ID token and refreesh key are now stored in cookies. usergroup is unknown where its bieng stored. ToDo: Fix navigation on succesful login, and check if session persists. currently after refreshing the page it seems the authtoken disapears for a second before reappearing which might be messnig up the check on login.